+++ genvalidity-property/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+* Gave `genGeneratesValid` and `genGeneratesInvalid` much nicer output.
 * Removed nonsense shrinking from `genGeneratesValid` and `genGeneratesInvalid`.
 
 ## [0.4.0.0] - 2019-03-08